requirements.txt
neat-python
pygame
Pillow
# pip install -r requirements.txt